+++ CHANGELOG.md
@@ -9,7 +9,10 @@ Versioning](http://semver.org/spec/v2.0.0.html) except to the first release.
 ## [Unreleased]
 
 ### Added
-
+- Add err log to `ConnectionPool.Add()` in case, when unable to establish 
+  connection and ctx is not canceled; 
+  also added logs for error case of `ConnectionPool.tryConnect()` calls in 
+  `ConnectionPool.controller()` and `ConnectionPool.reconnect()`
 ### Changed
 
 ### Fixed